+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 5.1.0
 -----
 
- * Added `allowEmptyValue` method to `NumericNodeDefinition` and `BooleanNodeDefinition` to accept `null` values.
+ * Added `allowNull` method to `NumericNodeDefinition` and `BooleanNodeDefinition` to accept `null` values.
 
 5.0.0
 -----